revert move WorkflowInputs and WorkflowSource to components folder
