+++ src/pages/SubmitWorkflow.js
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import { div, h, h2 } from 'react-hyperscript-helpers'
 import { ButtonOutline, ButtonPrimary, headerBar } from 'src/components/common'
-import { WorkflowInputs } from 'src/components/WorkflowInputs'
-import { WorkflowSource } from 'src/components/WorkflowSource'
+import { WorkflowInputs } from 'src/pages/WorkflowInputs'
+import { WorkflowSource } from 'src/pages/WorkflowSource'
 import { Ajax } from 'src/libs/ajax'
 import { useCancellation } from 'src/libs/react-utils'
